# Conflicts:
#	core/src/main/java/com/jjg/game/core/data/PropInfo.java
#	core/src/main/java/com/jjg/game/core/utils/PropUtil.java
#	gm/resources/sample/Item.xlsx
#	gm/resources/sample/global.xlsx
#	hall/resources/sample/Item.xlsx
#	hall/resources/sample/global.xlsx
#	hall/resources/sample/warehouse.xlsx
#	ploy/resources/sample/Item.xlsx
#	ploy/resources/sample/global.xlsx
#	ploy/resources/sample/warehouse.xlsx
#	poker/resources/sample/Item.xlsx
#	poker/resources/sample/global.xlsx
#	poker/resources/sample/warehouse.xlsx
#	recharge/resources/sample/Item.xlsx
#	slots/resources/sample/BaseRoller.xlsx
#	slots/resources/sample/Item.xlsx
#	slots/resources/sample/SpecialGird.xlsx
#	slots/resources/sample/SpecialMode.xlsx
#	slots/resources/sample/SpecialResultLib.xlsx
#	slots/resources/sample/global.xlsx
#	slots/resources/sample/warehouse.xlsx
#	slots/src/main/java/com/jjg/game/slots/data/SpecialResultLibCacheData.java
#	table/resources/sample/Item.xlsx
#	table/resources/sample/global.xlsx
#	table/resources/sample/warehouse.xlsx

+++ core/src/main/java/com/jjg/game/core/utils/PokerCardUtils.java
@@ -43,7 +43,7 @@ public class PokerCardUtils {
      * @param cardId 牌ID
      * @return 点数
      */
-    public static byte getPointId(byte cardId) {
+    public static byte getPointId(int cardId) {
         return (byte) (((cardId - 1) % 13) + 1);
     }
 
@@ -53,7 +53,7 @@ public class PokerCardUtils {
      * @param cardId 点数
      * @return 花色枚举
      */
-    public static EPokerSuit getSuit(byte cardId) {
+    public static EPokerSuit getSuit(int cardId) {
         return EPokerSuit.getPokerSuitById((cardId - 1) / 13 + 1);
     }
